avances sección 11

+++ TechNotes.Application/Notes/DeleteNote/DeleteNoteCommandHandler.cs
@@ -1,16 +1,27 @@
-﻿namespace TechNotes.Application.Notes.DeleteNote
+﻿using TechNotes.Application.Users;
+
+namespace TechNotes.Application.Notes.DeleteNote
 {
     public class DeleteNoteCommandHandler : ICommandHandler<DeleteNoteCommand>
     {
         private readonly INoteRepository _noteRepository;
+        private readonly IUserService _userService;
 
-        public DeleteNoteCommandHandler(INoteRepository noteRepository)
+        public DeleteNoteCommandHandler(INoteRepository noteRepository, IUserService userService)
         {
             _noteRepository = noteRepository;
+            _userService = userService;
         }
 
         public async Task<Result> Handle(DeleteNoteCommand request, CancellationToken cancellationToken)
         {
+            var currentUserCanDelete = await _userService.CurrentUserCanUpdateNoteAsync(request.Id);
+
+            if (!currentUserCanDelete)
+            {
+                return Result.Fail<NoteResponse?>("No tienes permiso para eliminar esta nota");
+            }
+
             var deleted = await _noteRepository.DeleteNoteAsync(request.Id);
 
             if (!deleted)
